# Conflicts:
#	src/main/resources/templates/rest/restDtl.html

+++ src/main/java/com/restaurant/repository/StarRepository.java
@@ -1,7 +1,9 @@
 package com.restaurant.repository;
 
+
+
 import com.restaurant.entity.Star;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StarRepository extends JpaRepository<Star,Long> {
+public interface StarRepository extends JpaRepository<Star, Integer> {
 }
